refactor(connect): restartEmu util
